requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==23.0.0
ics
requests    
